+++ src/hybran/__init__.py
@@ -2,3 +2,5 @@ from . import BLAST
 from . import CDHIT
 from . import MCL
 from . import addEggnogAnnotation
+
+__version__ = "1.3.0"